refactor: 抽取 CodeGen 业务编码生成工具类
统一封装 Redis INCR + 前缀 + 补零格式化逻辑，供产品/仓库/供应商/
客户模块新增时生成主键编码使用，消除 4 处重复的 generateXxxCode
私有方法。

+++ erp-server/erp-sales/src/main/java/com/qiheng/erp/sales/domain/customer/entity/Customer.java
@@ -1,7 +1,6 @@
 package com.qiheng.erp.sales.domain.customer.entity;
 
 import java.io.Serial;
-import java.math.BigDecimal;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.annotation.IdType;
@@ -57,7 +56,7 @@ public class Customer implements Serializable {
 
     @Schema(description = "信用额度")
     @TableField("credit_limit")
-    private BigDecimal creditLimit;
+    private Integer creditLimit;
 
     @Schema(description = "状态：1启用，0禁用")
     @TableField("status")
@@ -73,7 +72,7 @@ public class Customer implements Serializable {
 
     @Schema(description = "最后维护人ID")
     @TableField("updated_by_id")
-    private Long updatedById;
+    private long updatedById;
 
     @Schema(description = "最后维护人姓名")
     @TableField("updated_by_name")
